AY-91-New Navbar Design and adjusting some styling
