+++ web/index.html
@@ -6,7 +6,7 @@
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <title>Hands Off</title>
   </head>
-  <body class="h-full w-full bg-fuchsia-100">
+  <body class="h-full w-full bg-white ">
     <div id="root" class="h-full w-full max-w-screen p-0 m-0"></div>
     <script type="module" src="/src/main.tsx"></script>
   </body>
